Dont create rolebinding cnv-images if not exist

+++ internal/models/ocp_account.go
@@ -707,8 +707,6 @@ func (a *OcpAccountProvider) Request(serviceUuid string, cloud_selector map[stri
 			return
 		}
 
-		if err != nil {
-			log.Logger.Error("Error decrypting kubeconfig", "error", err)
 		rnew.OCPApiUrl = selectedCluster.ApiUrl
 		rnew.OcpClusterName = selectedCluster.Name
 
@@ -896,34 +894,42 @@ func (a *OcpAccountProvider) Request(serviceUuid string, cloud_selector map[stri
 			return
 		}
 
-		rb := &rbacv1.RoleBinding{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "allow-clone-" + namespaceName[:min(51, len(namespaceName))],
-				Namespace: "cnv-images",
-			},
-			Subjects: []rbacv1.Subject{
-				{
-					Kind:      "ServiceAccount",
-					Name:      "default",
-					Namespace: namespaceName,
+		// Look if namespace 'cnv-images' exists
+		if _, err := clientset.CoreV1().Namespaces().Get(context.TODO(), "cnv-images", metav1.GetOptions{}); err == nil {
+
+			rb := &rbacv1.RoleBinding{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "allow-clone-" + namespaceName[:min(51, len(namespaceName))],
+					Namespace: "cnv-images",
+					Labels: map[string]string{
+						"serviceUuid": serviceUuid,
+						"guid":        annotations["guid"],
+					},
 				},
-			},
-			RoleRef: rbacv1.RoleRef{
-				Kind:     "ClusterRole",
-				Name:     "datavolume-cloner",
-				APIGroup: "rbac.authorization.k8s.io",
-			},
-		}
-
-		_, err = clientset.RbacV1().RoleBindings("cnv-images").Create(context.TODO(), rb, metav1.CreateOptions{})
-		if err != nil {
-			log.Logger.Error("Error creating rolebinding on cnv-images", "error", err)
-
-			if err := clientset.CoreV1().Namespaces().Delete(context.TODO(), namespaceName, metav1.DeleteOptions{}); err != nil {
-				log.Logger.Error("Error cleaning up the namespace", "error", err)
+				Subjects: []rbacv1.Subject{
+					{
+						Kind:      "ServiceAccount",
+						Name:      "default",
+						Namespace: namespaceName,
+					},
+				},
+				RoleRef: rbacv1.RoleRef{
+					Kind:     "ClusterRole",
+					Name:     "datavolume-cloner",
+					APIGroup: "rbac.authorization.k8s.io",
+				},
 			}
-			rnew.SetStatus("error")
-			return
+
+			_, err = clientset.RbacV1().RoleBindings("cnv-images").Create(context.TODO(), rb, metav1.CreateOptions{})
+			if err != nil {
+				log.Logger.Error("Error creating rolebinding on cnv-images", "error", err)
+
+				if err := clientset.CoreV1().Namespaces().Delete(context.TODO(), namespaceName, metav1.DeleteOptions{}); err != nil {
+					log.Logger.Error("Error cleaning up the namespace", "error", err)
+				}
+				rnew.SetStatus("error")
+				return
+			}
 		}
 
 		secrets, err := clientset.CoreV1().Secrets(namespaceName).List(context.TODO(), metav1.ListOptions{})
